+++ 5-onboard-with-tap/trustedpartnerlist.txt
@@ -1,1 +1,2 @@
 ﻿*.fawltytowers2.com
+*.microsoft.com
